#132 Logic responsible for authentification was moved from user.service to authentification.service
