+++ Cooper/ClientApp/src.v2/app/layouts/login/login.layout.html
@@ -10,6 +10,7 @@
           <div class = "col-md-8 centralContentContainer">
             <div>
                 <coop-login
+                [failedLogin]="failedLogin"
                 (signIn)="signIn($event)" 
                 (socialSignIn)="socialSignIn($event)"
                 ></coop-login>
