+++ Cooper/ClientApp/src.v2/app/layouts/login/login.layout.ts
@@ -1,10 +1,11 @@
 import { Component, OnInit, ViewEncapsulation, Input } from '@angular/core';
-import { UserService } from '@services/model-services/user.service';
+import { AuthentificationService } from '@services/authentification/authentification.service';
 import { MatInputModule} from '@angular/material';
 import {MatFormFieldModule} from '@angular/material/form-field';
 import {trigger, transition, style, query, group, animateChild, animate, keyframes, state} from '@angular/animations';
 // import { fader } from '../../../animations/route-animation';
 import { NgForm } from '@angular/forms';
+import {ActivatedRoute} from '@angular/router';
 
 @Component({
   selector: 'coop-login-layout',
@@ -13,16 +14,26 @@ import { NgForm } from '@angular/forms';
 })
 export class LoginLayoutComponent implements OnInit {
 
-  constructor(private userService: UserService) {
-    this.userService.checkAuthentification();
+  failedLogin = false;
+
+  constructor(private authService: AuthentificationService, private route: ActivatedRoute) {
+
+    this.route.params.subscribe(params => {
+      console.log(params);
+      if (params['failedLogin']) {
+        this.failedLogin = params['failedLogin'];
+      }
+    });
+
+    this.authService.checkAuthentification();
   }
 
   signIn(form: NgForm): void {
-    this.userService.login(JSON.stringify(form.value));
+    this.authService.login(JSON.stringify(form.value));
   }
 
   socialSignIn(platform: string): void {
-    this.userService.socialSignIn(platform);
+    this.authService.socialSignIn(platform);
   }
 
   ngOnInit() {
