+++ Cooper/ClientApp/src.v2/app/services/authentification/authentification.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { AuthentificationService } from './authentification.service';
+
+describe('AuthentificationService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: AuthentificationService = TestBed.get(AuthentificationService);
+    expect(service).toBeTruthy();
+  });
+});
